Restructured signup packages in interface adapters for clarity
