+++ src/main/Java/interface_adapter/signup/choose_account_type/AccountTypeState.java
@@ -1,4 +1,4 @@
-package interface_adapter.choose_account_type;
+package interface_adapter.signup.choose_account_type;
 
 public class AccountTypeState {
     // Account type?
