+++ src/main/Java/interface_adapter/signup/general_user_signup/UserSignupPresenter.java
@@ -1,4 +1,4 @@
-package interface_adapter.general_user_signup;
+package interface_adapter.signup.general_user_signup;
 
 import interface_adapter.ViewManagerModel;
 import use_case.signup.general_user_signup.UserSignupOutputBoundary;
